pyproject.toml
[tool.poetry]
name = "agro-opt"
version = "0.1.0"
description = "Data-driven agricultural crop yield prediction and recommendation system"
authors = ["AgroOpt Team"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"

# Data science core
pandas = "^2.2.0"
numpy = "^1.26.0"
scikit-learn = "^1.4.0"
scipy = "^1.12.0"

# ML models
xgboost = "^2.0.0"
lightgbm = "^4.3.0"

# MLOps
mlflow = "^2.11.0"

# API
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"
httpx = "^0.27.0"

# Frontend
streamlit = "^1.32.0"
plotly = "^5.20.0"

# Utilities
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
joblib = "^1.3.0"
pyyaml = "^6.0.1"
requests = "^2.31.0"

# Notebook
jupyter = "^1.0.0"
ipykernel = "^6.29.0"
matplotlib = "^3.8.0"
seaborn = "^0.13.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
flake8-bugbear = "^24.2.6"
pre-commit = "^3.6.0"
pytest = "^8.1.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"
mypy = "^1.9.0"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
